+++ src/stuart-ml/clustering/VectorWithNorm.lua
@@ -9,12 +9,10 @@ local Vectors = require 'stuart-ml.linalg.Vectors'
 local VectorWithNorm = class('VectorWithNorm')
 
 function VectorWithNorm:initialize(arg1, norm)
-  if not arg1.isInstanceOf or not arg1:isInstanceOf(Vector) then
-    -- arg1 is an array
-    self.vector = Vectors.dense(arg1)
-  else
-    -- arg1 is a Vector
+  if arg1.isInstanceOf and arg1:isInstanceOf(Vector) then
     self.vector = arg1
+  else -- arg1 is an array
+    self.vector = Vectors.dense(arg1)
   end
   self.norm = norm or 2.0
 end